requirements.txt
affine==2.4.0
attrs==26.1.0
certifi==2026.2.25
click==8.3.2
click-plugins==1.1.1.2
cligj==0.7.2
contourpy==1.3.3
cycler==0.12.1
fiona==1.10.1
fonttools==4.62.1
geopandas==1.1.3
ImageIO==2.37.3
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
matplotlib==3.10.8
networkx==3.6.1
numpy==2.4.4
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==3.0.2
pillow==12.2.0
pyogrio==0.12.1
pyparsing==3.3.2
pyproj==3.7.2
python-dateutil==2.9.0.post0
rasterio==1.5.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
shapely==2.1.2
six==1.17.0
threadpoolctl==3.6.0
tifffile==2026.3.3
